Added arithmetic methods to inputs generated in the DSL.
