+++ lib/mb/sound.rb
@@ -78,6 +78,9 @@ module MB
   end
 end
 
+require_relative 'sound/arithmetic_mixin'
+require_relative 'sound/io_sample_mixin'
+
 require_relative 'sound/io_base'
 require_relative 'sound/io_input'
 require_relative 'sound/io_output'
