+++ lib/mb/sound/arithmetic_mixin.rb
@@ -1,0 +1,25 @@
+module MB
+  module Sound
+    # Adds methods to any
+    module ArithmeticMixin
+      # Creates a mixer that adds this mixer's output to +other+.  Part of a
+      # DSL experiment for building up a signal graph.
+      def +(other)
+        Mixer.new([self, other])
+      end
+
+      # Creates a mixer that subtracts +other+ from this mixer's output.  Part
+      # of a DSL experiment for building up a signal graph.
+      def -(other)
+        Mixer.new([self, [other, -1]])
+      end
+
+      # Creates a multiplier that multiplies +other+ by this mixer's output.
+      # Part of a DSL experiment for building up a signal graph.
+      def *(other)
+        other.or_at(1) if other.is_a?(Tone) # Keep amplitude high if multiplying tones
+        Multiplier.new([self, other])
+      end
+    end
+  end
+end
